% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanvasXpressBio.R
\docType{package}
\name{CanvasXpressBio-package}
\alias{CanvasXpressBio}
\alias{CanvasXpressBio-package}
\title{CanvasXpress visualizations for Bioconductor data structures}
\description{
Bridges the CanvasXpress interactive visualization library to core
Bioconductor classes. The [cxplot()] generic renders a
[SummarizedExperiment::SummarizedExperiment] (and any class that extends it,
such as `SingleCellExperiment` or `DESeqDataSet`) as an interactive
CanvasXpress heatmap carrying its row and column annotations, and
[cxvolcano()] draws a volcano plot from a differential-expression result
table.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neuhausi/canvasXpress}
  \item Report bugs at \url{https://github.com/neuhausi/canvasXpress/issues}
}

}
\author{
\strong{Maintainer}: Isaac Neuhaus \email{imnphd@gmail.com}

}
\keyword{internal}
