% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanvasXpressBio.R
\docType{package}
\name{CanvasXpressBio-package}
\alias{CanvasXpressBio}
\alias{CanvasXpressBio-package}
\title{CanvasXpress visualizations for Bioconductor data structures}
\description{
Bridges the CanvasXpress interactive visualization library to core
Bioconductor classes. [cxplot()] renders a
[SummarizedExperiment::SummarizedExperiment] (and any class that extends it,
such as `SingleCellExperiment` or `DESeqDataSet`) as an interactive
CanvasXpress heatmap carrying its row and column annotations; [cxvolcano()]
draws a volcano plot from a differential-expression table; [cxsurvival()]
draws a Kaplan-Meier curve from `colData`; and [cxboxplot()] draws a paired
boxplot of one feature with per-subject connections.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neuhausi/canvasXpress}
  \item Report bugs at \url{https://github.com/neuhausi/canvasXpress/issues}
}

}
\author{
\strong{Maintainer}: Isaac Neuhaus \email{imnphd@gmail.com} (\href{https://orcid.org/0000-0002-5622-8683}{ORCID})

}
\keyword{internal}
